Require EventvariableProducer to accept VInputTags instead of InputTags
This is a HACK to quickly upgrade to CMSSW_15_0_10. The CMSSW commit
e7ecea1 added type checking to cms.PSet. Our framework treats the
"eventvariables" collection as both a VInputTag and InputTag, depending
on the producer being configured.

This warrants a larger rework on how we configure our producers.
But for now, make all producers use VInputTag so we can start
processing data.