Update to CMSSW_15_0_10 (#283)
* Update CMSSW version code

* Remove pattern from classes_def.xml

The pattern parameter has been removed in favor of
explicitly stating all classes that need to be added
to the event.

* Update tensorflow session to use options struct

The createSession() function was changed to use an options struct
instead directly passing the number of threads in CMSSW
commit 8a546bc.

* Remove OriginalFormatProducer

This producer has removed methods as of CMSSW_15_0_10. It isn't
used and so putting effort into fixing it would be not worth the
time.

* Require EventvariableProducer to accept VInputTags instead of InputTags

This is a HACK to quickly upgrade to CMSSW_15_0_10. The CMSSW commit
e7ecea1 added type checking to cms.PSet. Our framework treats the
"eventvariables" collection as both a VInputTag and InputTag, depending
on the producer being configured.

This warrants a larger rework on how we configure our producers.
But for now, make all producers use VInputTag so we can start
processing data.

* Add more types to Collections classes_def.xml

* Expand wildcard classes_def.xml into explicit definitions